Refactor repository structure and update GetByIdAsync return type
Moved repository classes out of the DatabaseContext folder to a top-level Repositories folder for better organization. Changed GetByIdAsync in IGenericRepository and its implementation to return nullable types. Updated DeleteLeaveTypeCommandHandler logic to handle null values correctly. Changed project output type from Exe to Library.

+++ src/Core/HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -13,13 +13,16 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
-        await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
-
-        if (leaveTypeToDelete == null)
+        if (leaveTypeToDelete != null)
         {
-            throw new NotFoundException(nameof(LeaveType), request.Id);
+            await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
+
+            if (leaveTypeToDelete == null)
+            {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
         }
-        
+
         return Unit.Value;
     }
 }
